+++ packages/ui/src/components/Icon/index.ts
@@ -1,1 +1,2 @@
-export * from './Icon';
+export { Icon, IconFontProvider } from './Icon';
+export type { IconProps } from './Icon';